requirements.txt
nltk==3.9.1
Pillow==11.1.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
matplotlib==3.10.1
